extract getFormatLegalities
